BarChart & Axis Enhancements

+++ src/HydraDoc/Elements/SVGElements.cs
@@ -321,7 +321,7 @@ namespace HydraDoc.Elements
 
         public SVGLine() { }
 
-        public SVGLine( int x1, int y1, int x2, int y2 )
+        public SVGLine( double x1, double y1, double x2, double y2 )
         {
             X1 = x1; Y1 = y1; X2 = x2; Y2 = y2;
         }
